token error handling

+++ helpers/Auth.php
@@ -6,10 +6,11 @@
         {
             $token = Auth::getToken();
             
-            if (Auth::isValidToken($token)) {
-                $userID = DB::db()->query("SELECT user_id FROM tokens WHERE token = :token", array(':token' => $token))[0]['user_id'];
-                return DB::db()->query("SELECT * FROM users WHERE id = :id", array(':id' => $userID))[0];
+            if (!isset($token)) {
+                return;
             }
+            $userID = DB::db()->query("SELECT user_id FROM tokens WHERE token = :token", array(':token' => $token))[0]['user_id'];
+            return DB::db()->query("SELECT * FROM users WHERE id = :id", array(':id' => $userID))[0];
         }
 
         public static function login($userID)
@@ -17,12 +18,19 @@
             $token = bin2hex(random_bytes(64));
             $timestamp = date("Y-m-d");
             DB::db()->query("INSERT INTO tokens (user_id, token, timestamp) VALUES (:user_id, :token, :timestamp)", array(':user_id' => $userID, ':token' => $token, ':timestamp' => $timestamp));
+            return $token;
         }
 
         public static function logout()
         {
             $token = Auth::getToken();
+            if (!isset($token)) {
+                return;
+            }
             $userID = Auth::user()['id'];
+            if (!isset($userID)) {
+                return;
+            }
             DB::db()->query("DELETE FROM tokens WHERE user_id = :user_id AND token = :token", array(':user_id' => $userID, ':token' => $token));
         }
 
@@ -33,8 +41,15 @@
                 $lastLogin = new DateTime($timestamp[0]['timestamp']);
                 $currentTime = new DateTime();
                 return $lastLogin->diff($currentTime)->d < 10;
+            } else {
+                http_response_code(401);
+                echo json_encode(
+                    array(
+                        "error" => "Invalid token"
+                    )
+                );
+                return;
             }
-            return false;
         }
 
         public static function userID()
@@ -42,7 +57,7 @@
             $token = Auth::getToken();
 
             if (isset($token)) {
-                return DB::db()->query("SELECT user_id FROM tokens WHERE token = :token", array(':token' => $token));
+                return DB::db()->query("SELECT user_id FROM tokens WHERE token = :token", array(':token' => $token))[0]['user_id'];
             }
         }
